+++ mex/backend/constants.py
@@ -1,1 +1,0 @@
-NUMBER_OF_RULE_TYPES = 3  # Additive, Subtractive and Preventive