+++ mex/backend/merged/helpers.py
@@ -2,7 +2,6 @@ from typing import Annotated, Any, cast
 
 from pydantic import Field, TypeAdapter, ValidationError
 
-from mex.backend.constants import NUMBER_OF_RULE_TYPES
 from mex.backend.fields import MERGEABLE_FIELDS_BY_CLASS_NAME
 from mex.backend.graph.connector import GraphConnector
 from mex.backend.graph.exceptions import InconsistentGraphError
@@ -153,23 +152,20 @@ def merge_search_result_item(item: dict[str, Any]) -> AnyMergedModel:
         for component in item["components"]
         if component["entityType"] in EXTRACTED_MODEL_CLASSES_BY_NAME
     ]
-    rules_raw = [
+    raw_rules = [
         component
         for component in item["components"]
         if component["entityType"] in RULE_MODEL_CLASSES_BY_NAME
     ]
-    if len(rules_raw) == NUMBER_OF_RULE_TYPES:
-        rule_set_response = transform_raw_rules_to_rule_set_response(rules_raw)
-    elif len(rules_raw) == 0:
-        rule_set_response = None
+    if raw_rules:
+        rule_set = transform_raw_rules_to_rule_set_response(raw_rules)
     else:
-        msg = f"Unexpected number of rules found in graph: {len(rules_raw)}"
-        raise InconsistentGraphError(msg)
+        rule_set = None
 
     return create_merged_item(
         identifier=item["identifier"],
         extracted_items=extracted_items,
-        rule_set=rule_set_response,
+        rule_set=rule_set,
     )