+++ tests/rules/test_helpers.py
@@ -9,7 +9,7 @@ from mex.common.types import TextLanguage
 @pytest.mark.parametrize(
     ("items", "expected"),
     [
-        ([], "inconsistent rule item count"),
+        ([], "inconsistent number of rules found: 0"),
         (
             [
                 {
@@ -42,7 +42,8 @@ from mex.common.types import TextLanguage
                     "stableTargetId": ["00000000000003"],
                 },
             ],
-            "inconsistent rule item stableTargetIds",
+            "inconsistent rule item stableTargetIds: "
+            "00000000000001, 00000000000002, 00000000000003",
         ),
         (
             [