+++ tests/rules/test_main.py
@@ -207,6 +207,17 @@ def test_create_rule_set(client_with_api_key_write_permission: TestClient) -> No
 
 
 @pytest.mark.integration
+def test_get_rule_set_not_found(
+    client_with_api_key_write_permission: TestClient,
+) -> None:
+    response = client_with_api_key_write_permission.get(
+        "/v0/rule-set/thisIdDoesNotExist"
+    )
+    assert response.status_code == status.HTTP_404_NOT_FOUND, response.text
+    assert response.json() == {"detail": "no rules found"}
+
+
+@pytest.mark.integration
 def test_get_rule_set(
     client_with_api_key_write_permission: TestClient,
     load_dummy_rule_set: OrganizationalUnitRuleSetResponse,